split picker
